Добавил логи + обработку
Логи основных действий;
Обработка ошибок;
Добавил дополнительные модели для удобства;

+++ DigDesNote/DigDesNote.API/Controllers/CategoryController.cs
@@ -1,12 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
 using System.Web.Http;
 using DigDesNote.Model;
 using DigDesNote.DataLayer;
 using DigDesNote.DataLayer.Sql;
+using System.ComponentModel.DataAnnotations;
+using DigDesNote.API.Models;
 
 namespace DigDesNote.API.Controllers
 {
@@ -17,6 +16,9 @@ namespace DigDesNote.API.Controllers
     {
         private String _connectionString = @"Data Source=DESKTOP-H4JQP0V;Initial Catalog=NoteDb;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=True;ApplicationIntent=ReadWrite;MultiSubnetFailover=False";
         private ICategoriesRepository _categoriesRepository;
+
+        private List<ValidationResult> result;
+        private ValidationContext context;
 
         public CategoryController()
         {
@@ -29,11 +31,16 @@ namespace DigDesNote.API.Controllers
         /// <param name="id">ID категории</param>
         /// <returns></returns>
         [HttpGet]
+        [CustomExceptionAtribute]
         [Route("api/category/{id}")]
         public Category GetCategoryInfo(Guid id)
         {
-            Logger.Log.Instance.Info($"Получить информацию о категории {id}");
-            return _categoriesRepository.Get(id);
+            Logger.Log.Instance.Info($"Попытка получить информацию о категории {id}");
+            Category category = _categoriesRepository.Get(id);
+
+            if (category == null) throw new NoFoundException($"Категория с id = {id} не найдена");
+            else return category;
+
         }
 
         /// <summary>
@@ -42,26 +49,44 @@ namespace DigDesNote.API.Controllers
         /// <param name="category"></param>
         /// <returns></returns>
         [HttpPost]
+        [CustomExceptionAtribute]
         [Route("api/category")]
-        public Category CreateCategory([FromBody]Category category)
+        public Category CreateCategory([FromBody]CategoryCreate category)
         {
-            Logger.Log.Instance.Info($"Пользователь {category._userId} создал категорю {category._name}");
-            return _categoriesRepository.Create(category);
+            Logger.Log.Instance.Info($"Попытка пользователя {category._userId} создать категорю {category._name}");
+
+            result = new List<ValidationResult>();
+            var context = new ValidationContext(category);
+
+            if (!Validator.TryValidateObject(category, context, result, true))
+            {
+                String message = "Не удалось создать категорию. Указаны не все параметры: ";
+                foreach (var error in result) message += Environment.NewLine + error.ErrorMessage;
+                throw new ModelNotValid(message);
+            }
+            else
+            {
+                Category cat = _categoriesRepository.Create(category._userId, category._name);
+                Logger.Log.Instance.Info($"Категория {category._name} была создана пользователем {cat._userId}. id = {cat._id}");
+                return cat;
+            }
         }
 
-        /// <summary>
-        /// Создать категорию
-        /// </summary>
-        /// <param name="name">Имя категории</param>
-        /// <param name="id">ID пользователя</param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("api/category/{name}/{id}")]
-        public Category CreateCategory(String name, Guid id)
-        {
-            Logger.Log.Instance.Info($"Пользователь {id} создал категорю {name}");
-            return _categoriesRepository.Create(id, name);
-        }
+        ///// <summary>
+        ///// Создать категорию
+        ///// </summary>
+        ///// <param name="name">Имя категории</param>
+        ///// <param name="id">ID пользователя</param>
+        ///// <returns></returns>
+        //[HttpPost]
+        //[Route("api/category/{name}/{id}")]
+        //public Category CreateCategory(String name, Guid id)
+        //{
+        //    Logger.Log.Instance.Info($"Попытка пользователя {id} создать категорю {name}");
+        //    Category cat = _categoriesRepository.Create(id, name);
+        //    Logger.Log.Instance.Info($"Категория {name} была создана пользователем {id} {Environment.NewLine} cat = {cat._id}");
+        //    return cat;
+        //}
 
         /// <summary>
         /// Обновить категорию
@@ -69,36 +94,56 @@ namespace DigDesNote.API.Controllers
         /// <param name="category"></param>
         /// <returns></returns>
         [HttpPut]
-        [Route("api/category/{id}")]
-        public Category UpdateCategory(Category category)
+        [CustomExceptionAtribute]
+        [Route("api/category")]
+        public Category UpdateCategory(CategoryRename category)
         {
             Logger.Log.Instance.Info($"Попытка переименовать категорию {category._id} в {category._name}");
-            return _categoriesRepository.Edit(category._id, category._name);
+
+            if (_categoriesRepository.Get(category._id) == null) throw new NoFoundException($"Категория {category._id} не найдена");
+
+            result = new List<ValidationResult>();
+            context = new ValidationContext(category);
+
+            if (!Validator.TryValidateObject(category, context, result, true))
+            {
+                String message = "Не удалось переименовать категорию. Указаны не все данные: ";
+                foreach (var error in result) message += Environment.NewLine + error.ErrorMessage;
+                throw new ModelNotValid(message);
+            }
+            else
+            {
+                Category cat = _categoriesRepository.Edit(category._id, category._name);
+                Logger.Log.Instance.Info($"Категория {cat._id} была переименована в {cat._name}");
+                return cat;
+            }
         }
 
-        /// <summary>
-        /// Обновить категорию
-        /// </summary>
-        /// <param name="id"></param>
-        /// <param name="newName"></param>
-        /// <returns></returns>
-        [HttpPut]
-        [Route("api/category/{id}/{newName}")]
-        public Category UpdateCategory(Guid id, String newName)
-        {
-            Logger.Log.Instance.Info($"Попытка переименовать категорию {id} в {newName}");
-            return _categoriesRepository.Edit(id, newName);
-        }
+        ///// <summary>
+        ///// Обновить категорию
+        ///// </summary>
+        ///// <param name="id"></param>
+        ///// <param name="newName"></param>
+        ///// <returns></returns>
+        //[HttpPut]
+        //[Route("api/category/{id}/{newName}")]
+        //public Category UpdateCategory(Guid id, String newName)
+        //{
+        //    Logger.Log.Instance.Info($"Попытка переименовать категорию {id} в {newName}");
+        //    return _categoriesRepository.Edit(id, newName);
+        //}
 
         /// <summary>
         /// Удалить категорию
         /// </summary>
         /// <param name="id"></param>
         [HttpDelete]
+        [CustomExceptionAtribute]
         [Route("api/category/{id}")]
         public void DeleteCategory(Guid id)
         {
             Logger.Log.Instance.Info($"Попытка удалить категорию {id}");
+            if (_categoriesRepository.Get(id) == null) throw new NoFoundException($"Категория {id} не найдена");
             _categoriesRepository.Delete(id);
             Logger.Log.Instance.Info($"Категория {id} была удалена");
         }
